add podspec filetype
- factor out code from XCconfigFile to extractVersionStringFromTextFile
- slightly improve logic when finding multiple keys in text file, to only
consider lines that contain a valid version definition
- introduce TextFile protocol that describes how to take a line from a file
and extract the version string from it
- expand use of generics on the File and Version protocols, but not completely
yet

Also:
- fix bug with replacing new version numbers, first noticed in the
new podspec tests. E.g. there is semantic version of 1.0.3 and a numeric
version of 1 defined in a TextFile type file like podspec or xcconfig.
When replacing a numeric version bumped
from 1 to 2, both versions were modified: 1 -> 2 and incorrectly,  1.0.3 ->
2.0.3. Can't simply search/replace on old version string.